Reimplemented the parser with a token-based system
It make it easier and more effecient to parse a configuration file,
the parser has been redesigned and implemented into a token-based
parser. The lexer takes a reader and tokenizes it before going to
the parser. Through this change the syntax as also been relaxed
slightly. Commas are no longer used as option value separators
however due to the implementation, it's not a backwards
compatibility break. Any existing comma will simply be passed over.
Another new addition is the ability to use double quotes around
strings for example the domain name. Double quotes cannot be used
around IP addresses or where a string would otherwise not be logical.
All files using the previous syntax are still valid with the new
parser. New files that take advantage of new syntax will of course
not be compatible with the old parser.