The parser now reads from the lexer one token at a time
Previously the entire byte slice had to be read into memory and
be processed in bulk by the lexer. Now it's read as needed by
request of the parser. A small buffer has been created so
only a single token can be returned for each call to next()
even though some consume functions return multiple tokens.