Closes #12 - Allow manual DHCP option codes
For options which are not defined in the package itself, an option
can be declared using the syntax "option option-xxx value" where
xxx = an int between 1 and 255. Value can be one or more of string,
ip addresses, integers, or boolean values. All values to the end
of the line are assumed to be part of the option. No validation is
done so the user is required to ensure the values are correct.