Convert to standalong DHCP server and use separate verification system
The server will connect to a verification server that will determine
if a client is allowed on a particular network or not. The protocol
is binary and will be documented separatly.

The Packet Guardian web server will similarly be redesigned to respond
to verification requests and it will be responsible for determining
if a client is "registered" or not.