Removed duplicate IP check, never disable stdout logging
The logging.disable setting now disables only file logging. Stdout
logging will always happen.

Added a TODO for the handleDecline method. It won't work according
to the RFC.

Removed the duplicate IP check. The host should do a check before
it starts to use an offered IP. If it's in use the client will send
a DHCP DECLINE message. The server doesn't need to check and in many
cases in probably can't check reliably.