Added device storage, added BlockBlacklisted setting
A management console will be responsible for setting device
information such as if a device is registered or blacklisted.
The default is unregistered and not blacklisted which would
make it run as a normal DHCP server. RPC calls to modify
devices will come later.

The BlockBlacklisted setting will make the server not respond
to blacklisted devices rather than treating them as unregistered.
By default this is false which means blacklisted will be treated
as unregistered.

Also fixed failing tests now that device management is back.