Create RPC management server, create RPC client package
The main changes are the addition of the management server
and a client library to use it. Currently on the Device
service works because I failed to read the documentation
the first time and didn't use the correct function interface
for the RPC methods. The cli command is a program to test
the RPC interface. Eventually it will be a full featured
management cli for the server along with the main web interface
from Packet Guardian.