Added PGStore, fixed prepared statement issue, added device test
And a few other misc items.

The PGStore is like the MySQLStore but with custom device methods
to deal with the database schema of Packet Guardian's main database.
This will allow the new server to be used in-place of the old server
with no modifications to the management server. Hopefully.

The prepared statements were going stale since they are tied to
a specific database connection. They are now part of the store
struct and setup as needed.

Added an extra device test to test getting a device that is blacklisted.
This ensures the PGStore is working properly since it deals with
two separate tables for registration and blacklist information.