Remove need for local network in configuration
The application loads the server interface and unicast addresses on
start and will automatically handle packets coming in directly via
IP broadcast rather than through a relay. The application takes the
first global unicast address of each interface and will use that
as the "gateway" address if the GI field is zero.

The configuration parser will still allow the "local" keyword, but
it no longer does anything.