Don't set giaddr if the packet is not broadcast
Fixes an issue where client request packets were being rejected
due to "Gateway not found". When a client was requesting a lease
renewal, which happens via unicast UDP, the packet giaddr was
being set since it was empty to the IP address of the server.
This is needed for the server logic for local network requests,
but unicast client requests were being altered mistakenly. This
caused the server to always send a NACK for a lease renewal forcing
the client to restart the DHCP cycle.