Refactored gateway cache to function
Discover, request, and decline now use the same logic to
retrieve the correct network based on the giaddr of an
incoming packet. This fixes a bug where the server would
respond with NACK if it restarts, thus invalidating the
gateway cache, and a client sents a REQUEST to renew its
lease. This forced the client to restart the whole DHCP
process. Now, all functions that need the gateway cache
can also search for and set the cache thus allowing existing
leases to be renewed normally.